Auditors are only used by the registration service, no need to live in common
