+++ registration/app/registration/auditor/AuditorApiConfig.scala
@@ -1,0 +1,5 @@
+package registration.auditor
+
+case class AuditorApiConfig(url: String, apiKey: String)
+
+case class AuditorGroupConfig(contentApiConfig: AuditorApiConfig, paApiConfig: AuditorApiConfig)
